requirements.txt
requests
minicli
PyYAML
